# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MOLOCHAIN API Documentation",
    "description": "API documentation for the MOLOCHAIN global logistics and commodity platform",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@molochain.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/services": {
      "get": {
        "tags": ["services"],
        "summary": "Get all services",
        "description": "Get a list of all available services",
        "operationId": "getServices",
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": ["services"],
        "summary": "Get service by ID",
        "description": "Get details of a specific service by ID",
        "operationId": "getServiceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Service ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/ecosystem/registry/services": {
      "get": {
        "tags": ["ecosystem-registry"],
        "summary": "List ecosystem services",
        "description": "Get all registered services in the Molochain ecosystem",
        "operationId": "listEcosystemServices",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "maintenance"]
            },
            "description": "Filter by service status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ecosystem services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EcosystemServicesResponse"
                }
              }
            }
          }
        },
        "security": [{ "cookieAuth": [] }]
      },
      "post": {
        "tags": ["ecosystem-registry"],
        "summary": "Register new service",
        "description": "Register a new service in the ecosystem (admin only)",
        "operationId": "registerEcosystemService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EcosystemServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Service registered successfully"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
...
    "/ecosystem/registry/webhooks": {
      "get": {
        "tags": ["ecosystem-registry"],
        "summary": "List webhooks",
        "description": "Get all registered webhook endpoints",
        "operationId": "listEcosystemWebhooks",
        "responses": {
          "200": {
            "description": "List of webhooks"
          }
        },
        "security": [{ "cookieAuth": [] }]
      },
      "post": {
        "tags": ["ecosystem-registry"],
        "summary": "Register webhook",
        "description": "Register a new webhook endpoint for event notifications",
        "operationId": "registerEcosystemWebhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EcosystemWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Webhook registered"
          }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Service ID",
            "example": "ocean-fcl"
          },
          "code": {
            "type": "string",
            "description": "Service code",
            "example": "OCE-FCL"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "Ocean Freight (FCL)"
          },
          "description": {
            "type": "string",
            "description": "Service description",
            "example": "Full container load ocean shipping"
          },
          "category": {
            "type": "string",
            "description": "Service category",
            "example": "ocean"
          },
          "active": {
            "type": "boolean",
            "description": "Whether the service is active",
            "example": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "INTERNAL_ERROR"
          }
        }
      },
...
      "EcosystemServicesResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "services": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EcosystemService" }
          }
        }
      },
      "EcosystemService": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "example": "RAYANAVA AI Gateway" },
          "slug": { "type": "string", "example": "rayanava" },
          "baseUrl": { "type": "string", "example": "http://127.0.0.1:5001" },
          "healthEndpoint": { "type": "string", "example": "/health" },
          "status": { "type": "string", "enum": ["active", "inactive", "maintenance"] },
          "lastHealthCheck": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "EcosystemServiceRequest": {
        "type": "object",
        "required": ["name", "slug", "baseUrl"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "baseUrl": { "type": "string" },
          "healthEndpoint": { "type": "string", "default": "/health" },
          "metadata": { "type": "object" }
        }
      },
      "EcosystemWebhookRequest": {
        "type": "object",
        "required": ["serviceId", "targetUrl", "events"],
        "properties": {
          "serviceId": { "type": "integer" },
          "targetUrl": { "type": "string", "format": "uri" },
          "events": { "type": "array", "items": { "type": "string" } },
          "secret": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}